refactor(#3038): extract submit_review_decision body into services::review_decision (S1)
Move the decoded review-decision orchestration boundary from src/server/routes/review_verdict/decision_route.rs into crate::services::review_decision. The route now only unpacks State/AppState + Json<ReviewDecisionBody> and delegates to the service after request decode, before decision_route_validate_input and before the first DB side effect. The service keeps the legacy DecisionResponse shape for this verbatim slice so every HTTP status/body mapping stays byte-for-byte equivalent; domainizing the response can be a later behavior-changing seam, not this move.

The prior #3294 phase modules move from server/routes/review_verdict/decision_route/ to services/review_decision/. review_state_repo and the tuning aggregation implementation move with them so services::review_decision has no crate::server backflow. A thin server/routes/review_verdict/tuning_aggregate.rs wrapper remains only for the /reviews/tuning/aggregate route inventory source resolver; it delegates to the service implementation. The only body path adjustment in the moved code is the already-reexported event/card adapter path: server::routes::kanban::load_card_json_pg -> db::kanban_cards::load_card_json_pg and server::ws::emit_event -> eventbus::emit_event. Other changes are module-depth imports and state threading (&state in the route-owned body becomes state because the service parameter is already &AppState).

Equivalence proof: normalized token comparison of the 10 moved leaf files (accept, adapters, dismiss_finalize, dispute, pending, repo_card, repo_dispatch, worktree_stale, review_state_repo, tuning_aggregate) is 19301 -> 19301 tokens, identical after only the allowed path/import normalizations above. The submit_review_decision orchestrator body is 284 -> 284 tokens, identical after normalizing &state -> state. No transaction boundaries, advisory-lock order, idempotency checks, early returns, or response JSON/status branches were reordered.

Accounting: decision_route.rs ratchet lowered 139 -> 26 prod LoC. services/review_decision.rs is 335 prod LoC; service child modules are all <= 746 prod LoC and ratcheted at their landed values. service_server_backflow remains 0. route_srp current count is 13 (baseline 15).

Gates run:

- cargo fmt --check
- cargo check --workspace --all-targets
- cargo clippy --workspace --all-targets --all-features -- -D warnings
- cargo test --lib review (77 passed)
- cargo test --all-targets review_decision -- --skip _pg --skip pg_ --skip postgres (29 passed)
- cargo test --all-targets review_verdict -- --skip _pg --skip pg_ --skip postgres (1 passed)
- python3 scripts/audit_maintainability.py --check
- python3 scripts/generate_inventory_docs.py
- python3 scripts/check_agent_maintenance_docs.py
