+++ src/services/review_decision/accept.rs
@@ -10,8 +10,6 @@ use serde_json::json;
 use crate::app_state::AppState;
 use crate::services::review_decision::ReviewDecisionBody;
 
-use super::super::review_state_repo::update_card_review_state;
-use super::super::tuning_aggregate::record_decision_tuning;
 use super::DecisionResponse;
 use super::adapters::{
     consume_pending_review_decision_or_response, emit_card_updated,
@@ -27,6 +25,8 @@ use super::repo_card::{
 use super::repo_dispatch::{
     dispatch_status_and_result_pg_first, mark_next_review_round_advance_pg_first,
 };
+use super::review_state_repo::update_card_review_state;
+use super::tuning_aggregate::record_decision_tuning;
 
 /// Phase 3a of `submit_review_decision`: the `accept` decision branch (#195
 /// rework dispatch + #246 direct-review-skip-rework + #483 terminal
